Functioning Score Multiplier! (Need Create Powerup)
We now have all of the logic functioning perfectly in the backend, and in the frontend interaction with all of the UI. When doublePoints is set to true, the onscreen indicator showing the multiplier will show 2x instead of 1x, points per kill will be doubled, and it will all count towards the end level score which is displayed alongside the inlevel score. This is implemented in every level of the game.