Fully Functional High Score Counter, Levels 1-3!
I have implemented the UI and backend and frontend for the high score counter for our game, which shows the current high score for that level on that level's pause screen. The high score is logged after you complete a level if your score is higher than the previous high score (initialized to 0 when you start the game). This is fully implemented for all 3 levels of the game.